First try for downloading IRs/NAMs.
